requirements.txt
boto3_type_annotations>=0.3.1
ffmpy==0.3.0
jsonschema==4.4.0
pyjwt==2.4.0
pymediainfo==5.1.0
pymongo[srv]==3.12.0
python-dotenv==0.19.0
requests==2.26.0
sentry-sdk==1.5.12
filetype==1.1.0
